fix EA FA, change memory addressing to not use usize

+++ src/gpu.zig
@@ -1,8 +1,8 @@
 const std = @import("std");
 const IERegister = @import("ie_register.zig").IERegister;
 
-pub const VRAM_BEGIN: usize = 0x8000;
-pub const VRAM_END: usize = 0x9FFF;
+pub const VRAM_BEGIN: u16 = 0x8000;
+pub const VRAM_END: u16 = 0x9FFF;
 // const VRAM_SIZE: usize = VRAM_END - VRAM_BEGIN + 1;
 
 const TilePixelValue = enum {
